M5589 Integration test performance improvements
