+++ molgenis-platform-integration-tests/src/test/java/org/molgenis/integrationtest/platform/PlatformIT.java
@@ -18,10 +18,8 @@ import org.molgenis.data.meta.MetaDataServiceImpl;
 import org.molgenis.data.meta.model.*;
 import org.molgenis.data.meta.model.Package;
 import org.molgenis.data.security.EntityTypePermission;
-import org.molgenis.data.staticentity.TestEntityStatic;
 import org.molgenis.data.support.AggregateQueryImpl;
 import org.molgenis.data.support.QueryImpl;
-import org.molgenis.data.util.EntityUtils;
 import org.molgenis.data.validation.MolgenisValidationException;
 import org.molgenis.i18n.LanguageService;
 import org.slf4j.Logger;
@@ -36,10 +34,8 @@ import org.springframework.test.context.testng.AbstractTestNGSpringContextTests;
 import org.springframework.transaction.annotation.Transactional;
 import org.testng.annotations.*;
 
-import java.text.ParseException;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Supplier;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
@@ -51,10 +47,8 @@ import static java.util.Collections.singletonList;
 import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toSet;
-import static java.util.stream.Stream.concat;
 import static java.util.stream.Stream.of;
 import static org.molgenis.data.EntityTestHarness.*;
-import static org.molgenis.data.RepositoryCapability.*;
 import static org.molgenis.data.i18n.model.L10nStringMetaData.L10N_STRING;
 import static org.molgenis.data.i18n.model.LanguageMetadata.LANGUAGE;
 import static org.molgenis.data.meta.model.AttributeMetadata.ATTRIBUTE_META_DATA;
@@ -62,8 +56,6 @@ import static org.molgenis.data.meta.model.EntityTypeMetadata.ENTITY_TYPE_META_D
 import static org.molgenis.data.meta.model.PackageMetadata.PACKAGE;
 import static org.molgenis.data.security.EntityTypePermission.READ;
 import static org.molgenis.data.security.EntityTypePermission.WRITE;
-import static org.molgenis.data.util.MolgenisDateFormat.parseInstant;
-import static org.molgenis.data.util.MolgenisDateFormat.parseLocalDate;
 import static org.molgenis.security.core.runas.RunAsSystemAspect.runAsSystem;
 import static org.testng.Assert.*;
 
@@ -565,289 +557,6 @@ public class PlatformIT extends AbstractTestNGSpringContextTests
 		assertEquals(retrieved.count(), 0);
 	}
 
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testFindAll()
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(5);
-		Stream<Entity> retrieved = dataService.findAll(entityTypeDynamic.getId());
-		assertEquals(retrieved.count(), entities.size());
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testFindAllTyped()
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(1);
-		Supplier<Stream<Entity>> retrieved = () -> dataService.findAll(entityTypeDynamic.getId(), Entity.class);
-		assertEquals(retrieved.get().count(), 1);
-		assertEquals(retrieved.get().iterator().next().getIdValue(), entities.get(0).getIdValue());
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testFindAllByIds()
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(5);
-		Stream<Object> ids = Stream.concat(entities.stream().map(Entity::getIdValue), of("bogus"));
-		Stream<Entity> retrieved = dataService.findAll(entityTypeDynamic.getId(), ids);
-		assertEquals(retrieved.count(), entities.size());
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testFindAllByIdsTyped()
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createStaticAndAdd(5);
-
-		Supplier<Stream<TestEntityStatic>> retrieved = () -> dataService.findAll(entityTypeStatic.getId(),
-				Stream.concat(entities.stream().map(Entity::getIdValue), of("bogus")), TestEntityStatic.class);
-		assertEquals(retrieved.get().count(), entities.size());
-		assertEquals(retrieved.get().iterator().next().getId(), entities.get(0).getIdValue());
-		assertEquals(retrieved.get().iterator().next().getIdValue(), entities.get(0).getIdValue());
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testFindAllStreamFetch()
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(5);
-		Stream<Object> ids = concat(entities.stream().map(Entity::getIdValue), of("bogus"));
-		Stream<Entity> retrieved = dataService.findAll(entityTypeDynamic.getId(), ids, new Fetch().field(ATTR_ID));
-		assertEquals(retrieved.count(), entities.size());
-	}
-
-	@DataProvider(name = "findQueryOperatorEq")
-	private static Object[][] findQueryOperatorEq() throws ParseException
-	{
-		return new Object[][] { { ATTR_ID, "1", singletonList(1) }, { ATTR_STRING, "string1", asList(0, 1, 2) },
-				{ ATTR_BOOL, true, asList(0, 2) }, { ATTR_DATE, parseLocalDate("2012-12-21"), asList(0, 1, 2) },
-				{ ATTR_DATETIME, parseInstant("1985-08-12T11:12:13+0500"), asList(0, 1, 2) },
-				{ ATTR_DECIMAL, 1.123, singletonList(1) },
-				{ ATTR_HTML, "<html>where is my head and where is my body</html>", singletonList(1) },
-				{ ATTR_HYPERLINK, "http://www.molgenis.org", asList(0, 1, 2) },
-				{ ATTR_LONG, 1000000L, singletonList(1) }, { ATTR_INT, 11, singletonList(1) },
-				{ ATTR_SCRIPT, "/bin/blaat/script.sh", asList(0, 1, 2) },
-				{ ATTR_EMAIL, "this.is@mail.address", asList(0, 1, 2) },
-				// null checks
-				{ ATTR_ID, null, emptyList() }, { ATTR_STRING, null, emptyList() }, { ATTR_BOOL, null, emptyList() },
-				{ ATTR_CATEGORICAL, null, emptyList() }, { ATTR_CATEGORICAL_MREF, null, emptyList() },
-				{ ATTR_DATE, null, emptyList() }, { ATTR_DATETIME, null, emptyList() },
-				{ ATTR_DECIMAL, null, emptyList() }, { ATTR_HTML, null, asList(0, 2) },
-				{ ATTR_HYPERLINK, null, emptyList() }, { ATTR_LONG, null, emptyList() },
-				{ ATTR_INT, 11, singletonList(1) }, { ATTR_SCRIPT, null, emptyList() },
-				{ ATTR_EMAIL, null, emptyList() }, { ATTR_XREF, null, emptyList() }, { ATTR_MREF, null, emptyList() } };
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, dataProvider = "findQueryOperatorEq")
-	public void testFindQueryOperatorEq(String attrName, Object value, List<Integer> expectedEntityIndices)
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(3);
-		Supplier<Stream<Entity>> found = () -> dataService.query(entityTypeDynamic.getId()).eq(attrName, value).findAll();
-		List<Entity> foundAsList = found.get().collect(toList());
-		assertEquals(foundAsList.size(), expectedEntityIndices.size());
-		for (int i = 0; i < expectedEntityIndices.size(); ++i)
-		{
-			assertTrue(EntityUtils.equals(foundAsList.get(i), entities.get(expectedEntityIndices.get(i))));
-		}
-	}
-
-	@DataProvider(name = "findQueryOperatorIn")
-	private static Object[][] findQueryOperatorIn()
-	{
-		return new Object[][] { { singletonList("-1"), emptyList() }, { asList("-1", "0"), singletonList(0) },
-				{ asList("0", "1"), asList(0, 1) }, { asList("1", "2"), singletonList(1) } };
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, dataProvider = "findQueryOperatorIn")
-	public void testFindQueryOperatorIn(List<String> ids, List<Integer> expectedEntityIndices)
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(2);
-		Supplier<Stream<Entity>> found = () -> dataService.query(entityTypeDynamic.getId()).in(ATTR_ID, ids).findAll();
-		List<Entity> foundAsList = found.get().collect(toList());
-		assertEquals(foundAsList.size(), expectedEntityIndices.size());
-		for (int i = 0; i < expectedEntityIndices.size(); ++i)
-		{
-			assertTrue(EntityUtils.equals(foundAsList.get(i), entities.get(expectedEntityIndices.get(i))));
-		}
-	}
-
-	@DataProvider(name = "findQueryOperatorLess")
-	private static Object[][] findQueryOperatorLess()
-	{
-		return new Object[][] { { 9, emptyList() }, { 10, emptyList() }, { 11, singletonList(0) }, { 12, asList(0, 1) },
-				{ 13, asList(0, 1, 2) } };
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, dataProvider = "findQueryOperatorLess")
-	public void testFindQueryOperatorLess(int value, List<Integer> expectedEntityIndices)
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(5);
-		Supplier<Stream<Entity>> found = () -> dataService.query(entityTypeDynamic.getId()).lt(ATTR_INT, value).findAll();
-		List<Entity> foundAsList = found.get().collect(toList());
-		assertEquals(foundAsList.size(), expectedEntityIndices.size());
-		for (int i = 0; i < expectedEntityIndices.size(); ++i)
-		{
-			assertTrue(EntityUtils.equals(foundAsList.get(i), entities.get(expectedEntityIndices.get(i))));
-		}
-	}
-
-	@DataProvider(name = "findQueryOperatorLessEqual")
-	private static Object[][] findQueryOperatorLessEqual()
-	{
-		return new Object[][] { { 9, emptyList() }, { 10, singletonList(0) }, { 11, asList(0, 1) },
-				{ 12, asList(0, 1, 2) }, { 13, asList(0, 1, 2, 3) } };
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, dataProvider = "findQueryOperatorLessEqual")
-	public void testFindQueryOperatorLessEqual(int value, List<Integer> expectedEntityIndices)
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(5);
-		Supplier<Stream<Entity>> found = () -> dataService.query(entityTypeDynamic.getId()).le(ATTR_INT, value).findAll();
-		List<Entity> foundAsList = found.get().collect(toList());
-		assertEquals(foundAsList.size(), expectedEntityIndices.size());
-		for (int i = 0; i < expectedEntityIndices.size(); ++i)
-		{
-			assertTrue(EntityUtils.equals(foundAsList.get(i), entities.get(expectedEntityIndices.get(i))));
-		}
-	}
-
-	@DataProvider(name = "findQueryOperatorGreater")
-	private static Object[][] findQueryOperatorGreater()
-	{
-		return new Object[][] { { 9, asList(0, 1, 2) }, { 10, asList(1, 2) }, { 11, singletonList(2) },
-				{ 12, emptyList() } };
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, dataProvider = "findQueryOperatorGreater")
-	public void testFindQueryOperatorGreater(int value, List<Integer> expectedEntityIndices)
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(3);
-		Supplier<Stream<Entity>> found = () -> dataService.query(entityTypeDynamic.getId()).gt(ATTR_INT, value).findAll();
-		List<Entity> foundAsList = found.get().collect(toList());
-		assertEquals(foundAsList.size(), expectedEntityIndices.size());
-		for (int i = 0; i < expectedEntityIndices.size(); ++i)
-		{
-			assertTrue(EntityUtils.equals(foundAsList.get(i), entities.get(expectedEntityIndices.get(i))));
-		}
-	}
-
-	@DataProvider(name = "findQueryOperatorGreaterEqual")
-	private static Object[][] findQueryOperatorGreaterEqual()
-	{
-		return new Object[][] { { 9, asList(0, 1, 2) }, { 10, asList(0, 1, 2) }, { 11, asList(1, 2) },
-				{ 12, singletonList(2) }, { 13, emptyList() } };
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, dataProvider = "findQueryOperatorGreaterEqual")
-	public void testFindQueryOperatorGreaterEqual(int value, List<Integer> expectedEntityIndices)
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(3);
-		Supplier<Stream<Entity>> found = () -> dataService.query(entityTypeDynamic.getId()).ge(ATTR_INT, value).findAll();
-		List<Entity> foundAsList = found.get().collect(toList());
-		assertEquals(foundAsList.size(), expectedEntityIndices.size());
-		for (int i = 0; i < expectedEntityIndices.size(); ++i)
-		{
-			assertTrue(EntityUtils.equals(foundAsList.get(i), entities.get(expectedEntityIndices.get(i))));
-		}
-	}
-
-	@DataProvider(name = "findQueryOperatorRange")
-	private static Object[][] findQueryOperatorRange()
-	{
-		return new Object[][] { { 0, 9, emptyList() }, { 0, 10, singletonList(0) }, { 10, 10, singletonList(0) },
-				{ 10, 11, asList(0, 1) }, { 10, 12, asList(0, 1, 2) }, { 12, 20, singletonList(2) } };
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, dataProvider = "findQueryOperatorRange")
-	public void testFindQueryOperatorRange(int low, int high, List<Integer> expectedEntityIndices)
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(3);
-		Supplier<Stream<Entity>> found = () -> dataService.query(entityTypeDynamic.getId()).rng(ATTR_INT, low, high).findAll();
-		List<Entity> foundAsList = found.get().collect(toList());
-		assertEquals(foundAsList.size(), expectedEntityIndices.size());
-		for (int i = 0; i < expectedEntityIndices.size(); ++i)
-		{
-			assertTrue(EntityUtils.equals(foundAsList.get(i), entities.get(expectedEntityIndices.get(i))));
-		}
-	}
-
-	@DataProvider(name = "findQueryOperatorLike")
-	private static Object[][] findQueryOperatorLike()
-	{
-		return new Object[][] { { "ring", asList(0, 1) }, { "Ring", emptyList() }, { "nomatch", emptyList() } };
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, dataProvider = "findQueryOperatorLike")
-	public void testFindQueryOperatorLike(String likeStr, List<Integer> expectedEntityIndices)
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(2);
-		Supplier<Stream<Entity>> found = () -> dataService.query(entityTypeDynamic.getId()).like(ATTR_STRING, likeStr).findAll();
-		List<Entity> foundAsList = found.get().collect(toList());
-		assertEquals(foundAsList.size(), expectedEntityIndices.size());
-		for (int i = 0; i < expectedEntityIndices.size(); ++i)
-		{
-			assertTrue(EntityUtils.equals(foundAsList.get(i), entities.get(expectedEntityIndices.get(i))));
-		}
-	}
-
-	@DataProvider(name = "findQueryOperatorNot")
-	private static Object[][] findQueryOperatorNot()
-	{
-		return new Object[][] { { 9, asList(0, 1, 2) }, { 10, asList(1, 2) }, { 11, asList(0, 2) },
-				{ 12, asList(0, 1) }, { 13, asList(0, 1, 2) } };
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, dataProvider = "findQueryOperatorNot")
-	public void testFindQueryOperatorNot(int value, List<Integer> expectedEntityIndices)
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(3);
-		Supplier<Stream<Entity>> found = () -> dataService.query(entityTypeDynamic.getId()).not().eq(ATTR_INT, value).findAll();
-		List<Entity> foundAsList = found.get().collect(toList());
-		assertEquals(foundAsList.size(), expectedEntityIndices.size());
-		for (int i = 0; i < expectedEntityIndices.size(); ++i)
-		{
-			assertTrue(EntityUtils.equals(foundAsList.get(i), entities.get(expectedEntityIndices.get(i))));
-		}
-	}
-
 	/**
 	 * Test used as a caching benchmark
 	 */
@@ -876,332 +585,6 @@ public class PlatformIT extends AbstractTestNGSpringContextTests
 				dataService.findOne(entityTypeDynamic.getId(), q3);
 			}
 		});
-	}
-
-	@DataProvider(name = "findQueryOperatorAnd")
-	private static Object[][] findQueryOperatorAnd()
-	{
-		return new Object[][] { { "string1", 10, singletonList(0) }, { "unknownString", 10, emptyList() },
-				{ "string1", -1, emptyList() }, { "unknownString", -1, emptyList() } };
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, dataProvider = "findQueryOperatorAnd")
-	public void testFindQueryOperatorAnd(String strValue, int value, List<Integer> expectedEntityIndices)
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(3);
-		Supplier<Stream<Entity>> found = () -> dataService.query(entityTypeDynamic.getId())
-														  .eq(ATTR_STRING, strValue)
-														  .and()
-														  .eq(ATTR_INT, value)
-														  .findAll();
-		List<Entity> foundAsList = found.get().collect(toList());
-		assertEquals(foundAsList.size(), expectedEntityIndices.size());
-		for (int i = 0; i < expectedEntityIndices.size(); ++i)
-		{
-			assertTrue(EntityUtils.equals(foundAsList.get(i), entities.get(expectedEntityIndices.get(i))));
-		}
-	}
-
-	@DataProvider(name = "findQueryOperatorOr")
-	private static Object[][] findQueryOperatorOr()
-	{
-		return new Object[][] { { "string1", 10, asList(0, 1, 2) }, { "unknownString", 10, singletonList(0) },
-				{ "string1", -1, asList(0, 1, 2) }, { "unknownString", -1, emptyList() } };
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, dataProvider = "findQueryOperatorOr")
-	public void testFindQueryOperatorOr(String strValue, int value, List<Integer> expectedEntityIndices)
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(3);
-		Supplier<Stream<Entity>> found = () -> dataService.query(entityTypeDynamic.getId())
-														  .eq(ATTR_STRING, strValue)
-														  .or()
-														  .eq(ATTR_INT, value)
-														  .findAll();
-		List<Entity> foundAsList = found.get().collect(toList());
-		assertEquals(foundAsList.size(), expectedEntityIndices.size());
-		for (int i = 0; i < expectedEntityIndices.size(); ++i)
-		{
-			assertTrue(EntityUtils.equals(foundAsList.get(i), entities.get(expectedEntityIndices.get(i))));
-		}
-	}
-
-	@DataProvider(name = "findQueryOperatorNested")
-	private static Object[][] findQueryOperatorNested()
-	{
-		return new Object[][] { { true, "string1", 10, asList(0, 2) }, { true, "unknownString", 10, singletonList(0) },
-				{ true, "string1", -1, asList(0, 2) }, { true, "unknownString", -1, emptyList() },
-				{ false, "string1", 10, singletonList(1) }, { false, "unknownString", 10, emptyList() },
-				{ false, "string1", -1, singletonList(1) }, { false, "unknownString", -1, emptyList() } };
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, dataProvider = "findQueryOperatorNested")
-	public void testFindQueryOperatorNested(boolean boolValue, String strValue, int value, List<Integer> expectedEntityIndices)
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(3);
-		Supplier<Stream<Entity>> found = () -> dataService.query(entityTypeDynamic.getId())
-														  .eq(ATTR_BOOL, boolValue)
-														  .and()
-														  .nest()
-														  .eq(ATTR_STRING, strValue)
-														  .or()
-														  .eq(ATTR_INT, value)
-														  .unnest()
-														  .findAll();
-		List<Entity> foundAsList = found.get().collect(toList());
-		assertEquals(foundAsList.size(), expectedEntityIndices.size());
-		for (int i = 0; i < expectedEntityIndices.size(); ++i)
-		{
-			assertTrue(EntityUtils.equals(foundAsList.get(i), entities.get(expectedEntityIndices.get(i))));
-		}
-	}
-
-	@DataProvider(name = "findQueryOperatorSearch")
-	private static Object[][] findQueryOperatorSearch()
-	{
-		return new Object[][] { { "body", singletonList(1) }, { "head", singletonList(1) },
-				{ "unknownString", emptyList() } };
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, dataProvider = "findQueryOperatorSearch")
-	public void testFindQueryOperatorSearch(String searchStr, List<Integer> expectedEntityIndices)
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createDynamicAndAdd(2);
-		Supplier<Stream<Entity>> found = () -> dataService.query(entityTypeDynamic.getId())
-														  .search(ATTR_HTML, searchStr)
-														  .findAll();
-		List<Entity> foundAsList = found.get().collect(toList());
-		assertEquals(foundAsList.size(), expectedEntityIndices.size());
-		for (int i = 0; i < expectedEntityIndices.size(); ++i)
-		{
-			assertTrue(EntityUtils.equals(foundAsList.get(i), entities.get(expectedEntityIndices.get(i))));
-		}
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testFindQueryLimit2_Offset2_sortOnInt()
-	{
-		populateUserPermissions();
-
-		List<Entity> testRefEntities = testHarness.createTestRefEntities(refEntityTypeDynamic, 6);
-		List<Entity> testEntities = testHarness.createTestEntities(entityTypeDynamic, 10, testRefEntities).collect(toList());
-		runAsSystem(() ->
-		{
-			dataService.add(refEntityTypeDynamic.getId(), testRefEntities.stream());
-			dataService.add(entityTypeDynamic.getId(), testEntities.stream());
-		});
-		waitForIndexToBeStable(refEntityTypeDynamic, indexService, LOG);
-		waitForIndexToBeStable(entityTypeDynamic, indexService, LOG);
-		Supplier<Stream<Entity>> found = () -> dataService.findAll(entityTypeDynamic.getId(),
-				new QueryImpl<>().pageSize(2).offset(2).sort(new Sort(ATTR_ID, Sort.Direction.DESC)));
-		List<Entity> foundAsList = found.get().collect(toList());
-		assertEquals(foundAsList.size(), 2);
-		assertTrue(EntityUtils.equals(foundAsList.get(0), testEntities.get(7)));
-		assertTrue(EntityUtils.equals(foundAsList.get(1), testEntities.get(6)));
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testFindQueryTypedStatic()
-	{
-		populateUserPermissions();
-
-		List<Entity> entities = createStaticAndAdd(5);
-		Supplier<Stream<TestEntityStatic>> found = () -> dataService.findAll(entityTypeStatic.getId(),
-				new QueryImpl<TestEntityStatic>().eq(ATTR_ID, entities.get(0).getIdValue()), TestEntityStatic.class);
-		assertEquals(found.get().count(), 1);
-		assertEquals(found.get().findFirst().get().getId(), entities.get(0).getIdValue());
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testFindOne()
-	{
-		populateUserPermissions();
-
-		Entity entity = createDynamicAndAdd(1).get(0);
-		assertNotNull(dataService.findOneById(entityTypeDynamic.getId(), entity.getIdValue()));
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testFindOneTypedStatic()
-	{
-		populateUserPermissions();
-
-		Entity entity = createStatic(1).findFirst().get();
-		dataService.add(entityTypeStatic.getId(), Stream.of(entity));
-		waitForIndexToBeStable(entityTypeStatic, indexService, LOG);
-		TestEntityStatic testEntityStatic = dataService.findOneById(entityTypeStatic.getId(), entity.getIdValue(),
-				TestEntityStatic.class);
-		assertNotNull(testEntityStatic);
-		assertEquals(testEntityStatic.getId(), entity.getIdValue());
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testFindOneFetch()
-	{
-		populateUserPermissions();
-
-		Entity entity = createDynamicAndAdd(1).get(0);
-		assertNotNull(dataService.findOneById(entityTypeDynamic.getId(), entity.getIdValue(), new Fetch().field(ATTR_ID)));
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testFindOneFetchTypedStatic()
-	{
-		populateUserPermissions();
-
-		Entity entity = createStatic(1).collect(toList()).get(0);
-		entity.set(ATTR_ID, "1");
-		entity.set(ATTR_STRING, "string1");
-		entity.set(ATTR_BOOL, true);
-
-		dataService.add(entityTypeStatic.getId(), Stream.of(entity));
-		waitForIndexToBeStable(entityTypeStatic, indexService, LOG);
-
-		TestEntityStatic testEntityStatic = dataService.findOneById(entityTypeStatic.getId(), entity.getIdValue(),
-				new Fetch().field(ATTR_ID), TestEntityStatic.class);
-		assertNotNull(testEntityStatic);
-		assertEquals(testEntityStatic.getIdValue(), entity.getIdValue());
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testFindOneQuery()
-	{
-		populateUserPermissions();
-
-		Entity entity = createDynamicAndAdd(1).get(0);
-		entity = dataService.findOne(entityTypeDynamic.getId(), new QueryImpl<>().eq(ATTR_ID, entity.getIdValue()));
-		assertNotNull(entity);
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testFindOneQueryTypedStatic()
-	{
-		populateUserPermissions();
-
-		Entity entity = createStaticAndAdd(1).get(0);
-		TestEntityStatic testEntityStatic = dataService.findOne(entityTypeStatic.getId(),
-				new QueryImpl<TestEntityStatic>().eq(ATTR_ID, entity.getIdValue()), TestEntityStatic.class);
-		assertNotNull(testEntityStatic);
-		assertEquals(testEntityStatic.getId(), entity.getIdValue());
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testGetCapabilities()
-	{
-		populateUserPermissions();
-
-		Set<RepositoryCapability> capabilities = dataService.getCapabilities(entityTypeDynamic.getId());
-		assertNotNull(capabilities);
-		assertTrue(capabilities.containsAll(asList(MANAGABLE, QUERYABLE, WRITABLE, VALIDATE_REFERENCE_CONSTRAINT)));
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testGetEntityType()
-	{
-		populateUserPermissions();
-
-		EntityType entityType = dataService.getEntityType(entityTypeDynamic.getId());
-		assertNotNull(entityType);
-		assertTrue(EntityUtils.equals(entityType, entityTypeDynamic));
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testGetEntityNames()
-	{
-		populateUserPermissions();
-
-		Stream<String> names = dataService.getEntityTypeIds();
-		assertNotNull(names);
-		assertTrue(names.anyMatch(entityTypeDynamic.getId()::equals));
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testGetMeta()
-	{
-		populateUserPermissions();
-
-		assertNotNull(dataService.getMeta());
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testGetKnownRepository()
-	{
-		populateUserPermissions();
-
-		Repository<Entity> repo = dataService.getRepository(entityTypeDynamic.getId());
-		assertNotNull(repo);
-		assertEquals(repo.getName(), entityTypeDynamic.getId());
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, expectedExceptions = UnknownEntityTypeException.class)
-	public void testGetUnknownRepository()
-	{
-		populateUserPermissions();
-
-		dataService.getRepository("bogus");
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testHasRepository()
-	{
-		populateUserPermissions();
-
-		assertTrue(dataService.hasRepository(entityTypeDynamic.getId()));
-		assertFalse(dataService.hasRepository("bogus"));
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testIterator()
-	{
-		populateUserPermissions();
-
-		assertNotNull(dataService.iterator());
-		StreamSupport.stream(dataService.spliterator(), false).forEach(repo -> LOG.info(repo.getName()));
-		Repository repo = dataService.getRepository(entityTypeDynamic.getId());
-
-		/*
-			Repository equals is not implemented. The repository from dataService
-			and from the dataService.getRepository are not the same instances.
-		*/
-		assertTrue(StreamSupport.stream(dataService.spliterator(), false).anyMatch(e -> repo.getName().equals(e.getName())));
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true, expectedExceptions = UnknownEntityTypeException.class)
-	public void testQuery()
-	{
-		populateUserPermissions();
-
-		assertNotNull(dataService.query(entityTypeDynamic.getId()));
-		dataService.query("bogus");
 	}
 
 	@WithMockUser(username = USERNAME)
@@ -1807,96 +1190,6 @@ public class PlatformIT extends AbstractTestNGSpringContextTests
 			all.forEach(e -> LOG.info(e.getEntityTypeId() + "." + e.getEntityId()));
 			waitForIndexToBeStable(entityTypeDynamic, indexService, LOG);
 		});
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testOneDimensionalAggregateQuery()
-	{
-		populateUserPermissions();
-
-		createDynamicAndAdd(6);
-
-		AggregateResult result = runAggregateQuery(ATTR_BOOL, null, null, new QueryImpl<>());
-
-		AggregateResult expectedResult = new AggregateResult(asList(singletonList(3L), singletonList(3L)), asList(0L, 1L), emptyList());
-		assertEquals(result, expectedResult);
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testTwoDimensionalAggregateQuery()
-	{
-		populateUserPermissions();
-
-		createDynamicAndAdd(6);
-
-		AggregateResult result = runAggregateQuery(ATTR_BOOL, ATTR_ENUM, null, new QueryImpl<>());
-
-		AggregateResult expectedResult = new AggregateResult(asList(asList(0L, 3L), asList(3L, 0L)), asList(0L, 1L),
-				asList("option1", "option2"));
-		assertEquals(result, expectedResult);
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testOneDimensionalDistinctAggregateQuery()
-	{
-		populateUserPermissions();
-
-		createDynamicAndAdd(6);
-
-		AggregateResult result = runAggregateQuery(ATTR_BOOL, null, ATTR_ENUM, new QueryImpl<>());
-
-		AggregateResult expectedResult = new AggregateResult(asList(singletonList(1L), singletonList(1L)), asList(0L, 1L), emptyList());
-		assertEquals(result, expectedResult);
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testTwoDimensionalDistinctAggregateQuery()
-	{
-		populateUserPermissions();
-
-		createDynamicAndAdd(6);
-
-		AggregateResult result = runAggregateQuery(ATTR_BOOL, ATTR_BOOL, ATTR_ENUM, new QueryImpl<>());
-
-		AggregateResult expectedResult = new AggregateResult(asList(asList(1L, 0L), asList(0L, 1L)), asList(0L, 1L),
-				asList(0L, 1L));
-		assertEquals(result, expectedResult);
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testAggregateQueryWithQuery()
-	{
-		populateUserPermissions();
-
-		createDynamicAndAdd(6);
-		Query<Entity> query = new QueryImpl<>().gt(ATTR_INT, 12); // last 3 entities
-
-		AggregateResult result = runAggregateQuery(ATTR_BOOL, ATTR_ENUM, null, query);
-
-		AggregateResult expectedResult = new AggregateResult(asList(asList(0L, 2L), asList(1L, 0L)), asList(0L, 1L),
-				asList("option1", "option2"));
-		assertEquals(result, expectedResult);
-	}
-
-	@WithMockUser(username = USERNAME)
-	@Test(singleThreaded = true)
-	public void testDistinctAggregateQueryWithQuery()
-	{
-		populateUserPermissions();
-
-		createDynamicAndAdd(6);
-		Query<Entity> query = new QueryImpl<>().gt(ATTR_INT, 12); // last 3 entities
-
-		AggregateResult result = runAggregateQuery(ATTR_BOOL, ATTR_ENUM, ATTR_ENUM, query);
-
-		AggregateResult expectedResult = new AggregateResult(asList(asList(0L, 1L), asList(1L, 0L)), asList(0L, 1L),
-				asList("option1", "option2"));
-		assertEquals(result, expectedResult);
 	}
 
 	@WithMockUser(username = USERNAME)
